requirements.txt
flask
flask-cors
pillow
numpy
fpdf
twilio